+++ auth/apps/frontend/src/components/home/HomePage2/index.tsx
